update structure and add measurement
